test: move parsing tests into colocated module file
